+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] - 2019-04-26
+
+### Added
+
+- Docstrings expanded to note the source schema name.
+
+### Fixed
+
+- Docstrings were being stripped due to a bad tsconfig setting.
+
 ## [1.0.4] - 2019-03-26
 
 ### Added