+++ CHANGELOG.md
@@ -5,10 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] - 2019-04-26
+## [1.0.5] - 2019-04-27
 
 ### Added
 
+- Complete test coverage for mappers.
 - Docstrings expanded to note the source schema name.
 
 ### Fixed
@@ -16,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Docstrings were being stripped due to a bad tsconfig setting.
 - Fixed a bug in Platform Times where a Trip ID would always be undefined.
 - Fixed `Trip` interface as `id` and `wheelchairAccess` are optional according to the schema.
+- Change `PlatformTimes` optional properties to non-optional as they're always defined.
 
 ## [1.0.4] - 2019-03-26
 
@@ -70,7 +72,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added support for querying platform locations.
 - Added typings for all returned results.
 
-[unreleased]: https://github.com/Codex-/metro-info/compare/v1.0.4...HEAD
+[unreleased]: https://github.com/Codex-/metro-info/compare/v1.0.5...HEAD
+[1.0.5]: https://github.com/Codex-/metro-info/compare/v1.0.4...v1.0.5
 [1.0.4]: https://github.com/Codex-/metro-info/compare/v1.0.3...v1.0.4
 [1.0.3]: https://github.com/Codex-/metro-info/compare/v1.0.2...v1.0.3
 [1.0.2]: https://github.com/Codex-/metro-info/compare/v1.0.1...v1.0.2